% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelGraph.R
\name{modelGraph}
\alias{modelGraph}
\alias{plot_modelgraph}
\title{Model a graph based on numeric dataset}
\usage{
plot_modelgraph(
  data,
  x,
  y,
  model_type = "lm",
  remove_na = TRUE,
  axis_title_color = "black",
  axis_title_size = 12,
  ...
)
}
\arguments{
\item{data}{tibble}

\item{x}{Numeric column}

\item{y}{Numeric column}

\item{model_type}{String for type of model function used}

\item{remove_na}{Boolean of whether to remove NA values}

\item{axis_title_color}{Axis title color}

\item{axis_title_size}{Axis title size}

\item{...}{Ellipsis for other other arguments that can be used in ggplot}
}
\value{
A ggplot graph which contains a model type of the given numeric tibble.
}
\description{
Model a graph based on numeric dataset
}
\examples{
plot_modelgraph(mtcars, cyl, mpg, model_type = "lm")
}
